Finished

+++ src/main/java/com/crossover/techtrial/dto/TopMemberDTO.java
@@ -19,7 +19,7 @@ public class TopMemberDTO {
   public TopMemberDTO(Long memberId,
       String name, 
       String email, 
-      Integer bookCount) {
+      Long bookCount) {
     this.name = name;
     this.email = email;
     this.memberId = memberId;
@@ -36,7 +36,7 @@ public class TopMemberDTO {
   
   private String email;
   
-  private Integer bookCount;
+  private Long bookCount;
 
   public Long getMemberId() {
     return memberId;
@@ -62,11 +62,11 @@ public class TopMemberDTO {
     this.email = email;
   }
 
-  public Integer getBookCount() {
+  public Long getBookCount() {
     return bookCount;
   }
 
-  public void setBookCount(Integer bookCount) {
+  public void setBookCount(Long bookCount) {
     this.bookCount = bookCount;
   }
 }
